lustre: uapi: add new changerec_type
The Lazy Size on MDT is causing the trusted.som xattr to be logged
in the changelog whenever a file is needed to update this xattr
data casued by file open/close or truncate operations.

The original patch landed to the OpenSFS branch fixes this problem
to avoid logging this xattr for every file. This introduces a new
changelog_rec_type that the mdc changelog code needs to be aware
of.

WC-bug-id: https://jira.whamcloud.com/browse/LU-11450
Lustre-commit: faf6f514c172 ("LU-11450 mdd: avoid logging trusted.som xattr in changelogs")
Signed-off-by: Qian Yingjin <qian@ddn.com>
Reviewed-on: https://review.whamcloud.com/33323
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Wang Shilong <wshilong@ddn.com>
Reviewed-by: John L. Hammond <jhammond@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>