lnet: drop refcounting from event queues.
The refcount is no longer incremented or decremented
as there is no interesting state and no long term
references.  Nothing checks if LNetEQFree returns -EBUSY
any more - though a few calls LASSERT() that the return
value is 0.

So discard the refcounting, change LNetEQFree to return
'void', and remove all the LASSERTs.

WC-bug-id: https://jira.whamcloud.com/browse/LU-13005
Lustre-commit: c927c62f93feb ("LU-13005 lnet: drop refcounting from event queues.")
Signed-off-by: Mr NeilBrown <neilb@suse.de>
Reviewed-on: https://review.whamcloud.com/36846
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Reviewed-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Reviewed-by: Amir Shehata <ashehata@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>